Started on ship
JUst an idea of how it should work. Within the main game I believe we
should have the different ship types. We will tell the user to place the
specific ship and save the locations that they chose. The location array
in Ship might change to an array of tiles from grid however.